Revert "Add new HiddenFilter to filter annotations which are moderated or belong to a nipsaed user."
