+++ h/search/query.py
@@ -392,18 +392,19 @@ class DeletedFilter(object):
         return search.exclude("exists", field="deleted")
 
 
-class HiddenFilter(object):
-    """Return an Elasticsearch filter for filtering out moderated or NIPSA'd annotations."""
+class NipsaFilter(object):
+    """Return an Elasticsearch filter for filtering out NIPSA'd annotations."""
 
     def __init__(self, request):
         self.group_service = request.find_service(name='group')
         self.user = request.user
 
     def __call__(self, search, _):
-        """Filter out all hidden and NIPSA'd annotations except the current user's."""
-        # If any one of the "should" clauses is true then the annotation will
-        # get through these filter.
-        should_clauses = [Q("bool", must_not=Q("term", nipsa=True), must=Q("term", hidden=False))]
+        """Filter out all NIPSA'd annotations except the current user's."""
+        # If any one of these "should" clauses is true then the annotation will
+        # get through the filter.
+        should_clauses = [Q("bool", must_not=Q("term", nipsa=True)),
+                          Q("exists", field="thread_ids")]
 
         if self.user is not None:
             # Always show the logged-in user's annotations even if they have nipsa.
