+++ tests/h/search/conftest.py
@@ -6,7 +6,6 @@ import pytest
 
 import h.search.index
 from h.services.group import GroupService
-from h.services.annotation_moderation import AnnotationModerationService
 
 
 @pytest.fixture(autouse=True)
@@ -15,15 +14,6 @@ def group_service(pyramid_config):
     group_service.groupids_readable_by.return_value = ["__world__"]
     pyramid_config.register_service(group_service, name="group")
     return group_service
-
-
-@pytest.fixture(autouse=True)
-def moderation_service(pyramid_config):
-    svc = mock.create_autospec(AnnotationModerationService, spec_set=True, instance=True)
-    svc.all_hidden.return_value = []
-    svc.hidden.return_value = False
-    pyramid_config.register_service(svc, name='annotation_moderation')
-    return svc
 
 
 @pytest.fixture
